refactor: split host_boundary.go — extract overlay to host_overlay.go (1157→1157 lines)
Move overlay-related code (selfhostSpanIndex, overlaySelfhostResult,
buildSelfhostSpanIndex, addNode, lookup* methods, typeReprToType, etc.)
from host_boundary.go (2551→1394 lines) to new host_overlay.go.

The overlay remains opt-in via PopulateLegacyMaps flag.
This split isolates the legacy pointer-keyed map population from the
core checker dispatch, diagnostics, and caching infrastructure.
